+++ src/FreeElephants/RestDaemon/Endpoint/EndpointInterface.php
@@ -2,6 +2,7 @@
 
 namespace FreeElephants\RestDaemon\Endpoint;
 
+use FreeElephants\RestDaemon\Endpoint\Handler\EndpointMethodHandlerInterface;
 use FreeElephants\RestDaemon\Module\ApiModuleInterface;
 
 /**