+++ src/FreeElephants/RestDaemon/Endpoint/Handler/CloningHandlerFactory.php
@@ -2,13 +2,12 @@
 
 namespace FreeElephants\RestDaemon\Endpoint\Handler;
 
-use FreeElephants\RestDaemon\Endpoint\EndpointMethodHandlerInterface;
 use Psr\Container\ContainerInterface;
 
 /**
  * @author samizdam <samizdam@inbox.ru>
  */
-class HandlerFactory implements HandlerFactoryInterface
+class CloningHandlerFactory implements HandlerFactoryInterface
 {
 
     /**
@@ -23,6 +22,6 @@ class HandlerFactory implements HandlerFactoryInterface
 
     public function buildHandler(string $className): EndpointMethodHandlerInterface
     {
-        return new $className;
+        return clone $this->di->get($className);
     }
 }