+++ src/FreeElephants/RestDaemon/Endpoint/Handler/DefaultHandlerFactory.php
@@ -1,0 +1,14 @@
+<?php
+
+
+namespace FreeElephants\RestDaemon\Endpoint\Handler;
+
+
+class DefaultHandlerFactory implements HandlerFactoryInterface
+{
+
+    public function buildHandler(string $className): EndpointMethodHandlerInterface
+    {
+        return new $className;
+    }
+}